linktic_google_drive_api/requirements.txt
google_api_python_client==2.86.0
google_auth_oauthlib==1.0.0
protobuf==4.22.3
